+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 2.0.2
 -----
 - Support Pepe replacement on mobile.twitter.com.
+- Added a white index finger as a Pepe.
 
 2.0.1
 -----